feat: support cloning assets to created releases
