+++ gha_clone_releases/main.py
@@ -4,9 +4,9 @@ from typing import Any
 from actions_toolkit import core as actions_toolkit
 from github import Github
 from github.GithubException import GithubException
-from github.GitRelease import GitRelease
-from packaging.version import InvalidVersion
-from packaging.version import parse as parse_version
+
+from .utils.assets import download_asset
+from .utils.releases import get_missing_releases
 
 # This code is automatically generated by actions.yml and make generate-inputs
 ###START_INPUT_AUTOMATION###
@@ -20,6 +20,10 @@ INPUTS = {
     },
     "skip_draft": {"description": "Skip draft releases", "default": False},
     "skip_prerelease": {"description": "Skip Prereleases", "default": False},
+    "copy_assets": {
+        "description": "If true, copy assets from source repo releases to the newly created releases",
+        "default": False,
+    },
     "limit": {
         "description": "A limit of how many releases to add on a single run. Good for not overwhelming CI systems",
         "default": 0,
@@ -79,52 +83,17 @@ def get_inputs() -> dict[str, Any]:
         if isinstance(parsed_inputs["skip_prerelease"], str)
         else parsed_inputs["skip_prerelease"]
     )
+    parsed_inputs["copy_assets"] = (
+        parsed_inputs["copy_assets"].lower() == "true"
+        if isinstance(parsed_inputs["copy_assets"], str)
+        else parsed_inputs["copy_assets"]
+    )
     parsed_inputs["dest_repo"] = (
         os.environ.get("GITHUB_REPOSITORY") if parsed_inputs["dest_repo"] is None else parsed_inputs["dest_repo"]
     )
     if parsed_inputs["dest_repo"] is None:
         actions_toolkit.set_failed("Dest repo is none, set either INPUT_DEST_REPO or GITHUB_REPOSITORY")
     return parsed_inputs
-
-
-class ReleaseWrapper:
-    def __init__(self, release: GitRelease):
-        self.release = release
-
-    def __eq__(self, other_release: GitRelease):
-        return self.release.title == other_release.release.title
-
-    def __hash__(self):
-        return hash(self.release.title)
-
-
-def exceeds_min_version(release_version: str, min_version: str):
-    """Checks if a version exceeds a minimum version using packaging.version.parse"""
-    if min_version is not None:
-        try:
-            result = parse_version(release_version) > parse_version(min_version)
-        except InvalidVersion as exc:
-            actions_toolkit.error(f"{release_version} Is an invalid version {exc}")
-            return False
-        actions_toolkit.debug(f"{release_version} exceeds_min_version result {result}")
-        return result
-    return True
-
-
-def get_missing_releases(source_releases, dest_releases, min_version) -> list[GitRelease]:
-    """Compares two sets of releases and returns a list of releases in the source that are not in the destination"""
-
-    def sort_key(release: GitRelease):
-        return release.title
-
-    source_releases = {
-        ReleaseWrapper(release) for release in source_releases if exceeds_min_version(release.title, min_version)
-    }
-    dest_releases = {ReleaseWrapper(release) for release in dest_releases}
-
-    releases = [release.release for release in list(source_releases - dest_releases)]
-    releases.sort(key=lambda release: release.published_at, reverse=True)
-    return releases
 
 
 def main():
@@ -165,7 +134,7 @@ def main():
         actions_toolkit.info(f"Adding {release.tag_name} to {inputs['dest_repo']}")
         target = this_repo.default_branch if inputs["target"] is None else inputs["target"]
         try:
-            this_repo.create_git_release(
+            new_release = this_repo.create_git_release(
                 tag=release.tag_name,
                 name=release.title,
                 message=release.body,
@@ -176,6 +145,13 @@ def main():
             added_releases.append(release.tag_name)
         except GithubException as exc:
             actions_toolkit.error(f"Error while adding a release for {release.tag_name}. {exc}")
+        if inputs["copy_assets"]:
+            for asset in release.assets:
+                try:
+                    this_asset = download_asset(asset_name=asset.name, asset_url=asset.browser_download_url)
+                    new_release.upload_asset(this_asset)
+                except Exception as exc:
+                    actions_toolkit.error(f"Error while copying asset {asset.name}. {exc}")
     actions_toolkit.set_output("addedReleases", ",".join(added_releases))
     actions_toolkit.set_output("skippedReleasesCount", skipped)
     actions_toolkit.set_output("addedReleasesCount", len(added_releases))
